requirements.txt
opencv-python 
mediapipe 
tensorflow
fastapi
uvicorn
python-multipart
pillow